requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
anthropic
requests
python-dotenv
